parser/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
redis==5.0.1
loguru==0.7.2
